+++ src/algebra/SMatrix.h
@@ -26,8 +26,8 @@ class SerialMatrix {
   T* mat = nullptr;  // pointer to the internal array structure.
 
   /// Index from a 1D array to a position in a 2D array (matrix)
-  long global2Local(const long& row, const long& col);
-  std::tuple<long, long> local2Global(const long& k);
+  int global2Local(const int& row, const int& col);
+  std::tuple<int, int> local2Global(const int& k);
 
  public:
   /** Indicates that the matrix A is not modified: transN(A) = A
@@ -70,7 +70,7 @@ class SerialMatrix {
   /** Find the global indices of the matrix elements that are stored locally
    * by the current MPI process.
    */
-  std::vector<std::tuple<long, long>> getAllLocalStates();
+  std::vector<std::tuple<int, int>> getAllLocalStates();
 
   /** Returns true if the global indices (row,col) identify a matrix element
    * stored by the MPI process.
@@ -79,19 +79,19 @@ class SerialMatrix {
 
   /** Find global number of rows
    */
-  long rows() const;
+  int rows() const;
   /** Return local number of rows
   */
-  long localRows() const;
+  int localRows() const;
   /** Find global number of columns
    */
-  long cols() const;
+  int cols() const;
   /** Return local number of rows
   */
-  long localCols() const;
+  int localCols() const;
   /** Find global number of matrix elements
    */
-  long size() const;
+  int size() const;
   /** Get and set operator
    */
   T& operator()(const int &row, const int &col);
@@ -213,7 +213,7 @@ SerialMatrix<T>::SerialMatrix(const SerialMatrix<T>& that) {
   }
   mat = new T[numElements_];
   assert(mat != nullptr);
-  for (long i = 0; i < numElements_; i++) {
+  for (int i = 0; i < numElements_; i++) {
     mat[i] = that.mat[i];
   }
 }
@@ -230,7 +230,7 @@ SerialMatrix<T>& SerialMatrix<T>::operator=(const SerialMatrix<T>& that) {
     }
     mat = new T[numElements_];
     assert(mat != nullptr);
-    for (long i = 0; i < numElements_; i++) {
+    for (int i = 0; i < numElements_; i++) {
       mat[i] = that.mat[i];
     }
   }
@@ -245,23 +245,23 @@ SerialMatrix<T>::~SerialMatrix() {
 
 /* ------------- Very basic operations -------------- */
 template <typename T>
-long SerialMatrix<T>::rows() const {
+int SerialMatrix<T>::rows() const {
   return numRows_;
 }
 template <typename T>
-long SerialMatrix<T>::localRows() const {
+int SerialMatrix<T>::localRows() const {
   return numRows_;
 }
 template <typename T>
-long SerialMatrix<T>::cols() const {
+int SerialMatrix<T>::cols() const {
   return numCols_;
 }
 template <typename T>
-long SerialMatrix<T>::localCols() const {
+int SerialMatrix<T>::localCols() const {
   return numCols_;
 }
 template <typename T>
-long SerialMatrix<T>::size() const {
+int SerialMatrix<T>::size() const {
   return numElements_;
 }
 
@@ -290,7 +290,7 @@ bool SerialMatrix<T>::indicesAreLocal(const int& row, const int& col) {
 }
 
 template <typename T>
-std::tuple<long, long> SerialMatrix<T>::local2Global(const long& k) {
+std::tuple<int, int> SerialMatrix<T>::local2Global(const int& k) {
   // we convert this combined local index k into row / col indices
   // k = j * nRows + i
   if(numRows_ == 0) Error e("attempted to div by zero in l2g");
@@ -301,15 +301,15 @@ std::tuple<long, long> SerialMatrix<T>::local2Global(const long& k) {
 
 // Indexing to set up the matrix in col major format
 template <typename T>
-long SerialMatrix<T>::global2Local(const long& row, const long& col) {
+int SerialMatrix<T>::global2Local(const int& row, const int& col) {
   return numRows_ * col + row;
 }
 
 template <typename T>
-std::vector<std::tuple<long, long>> SerialMatrix<T>::getAllLocalStates() {
-  std::vector<std::tuple<long, long>> x;
-  for (long k = 0; k < numElements_; k++) {
-    std::tuple<long, long> t = local2Global(k);  // bloch indices
+std::vector<std::tuple<int, int>> SerialMatrix<T>::getAllLocalStates() {
+  std::vector<std::tuple<int, int>> x;
+  for (int k = 0; k < numElements_; k++) {
+    std::tuple<int, int> t = local2Global(k);  // bloch indices
     x.push_back(t);
   }
   return x;
